dang: generalize scalar API attachment beyond codecs
The JSON/YAML/TOML graft was gated on a codec-specific formatCodecs map and
named graftCodec*, reading as a carve-out for encoding. Make it a general
language mechanism: a builtin scalar is any ScalarKind static module Dang ships
members for, found via BuiltinScalarModule, and attachBuiltinSchemes /
attachBuiltinMethods staple those members onto any same-named scalar — owned,
user-declared, or schema-imported. Drop formatCodecs and move the mechanism to
builtins.go beside the rest of the builtin registry.

Behavior is unchanged: JSON/YAML/TOML still merge with a same-named schema or
user scalar, and a non-builtin scalar gets nothing. Encoding is now just the
first builtin scalar rather than a special case. Routing String's instance
methods through the same path is a larger eval-dispatch change, left for later.
